backend/requirements.txt
flask
flask-cors
werkzeug
google-generativeai
PyPDF2
python-docx
python-dotenv
textract
psycopg2-binary